Cripple SDL support for QEMU builds for SDK.
For PHA's uses, we DO NOT DO THIS.  It's not overly fast, to be honest
and doesn't gain you much that running in a stock QEMU (which is
available on your host or inside a Scratchbox mostly if not totally
the right ways...)- and it flatly doesn't compare to target, especially
if you're talking about anything other than framebuffer/X11 access
(If you're doing OpenGLES, OpenGL, Vulkan, etc. this is *WORTHLESS*.

Simply put, this is a Yocto-ism that isn't as useful as they lead
people to believe (much like the old SDK way of building in a QEMU
session for target...merp?).  Turning the *stupid* off in my
local.conf example.