Added new machine defaults to possible options.
NanoPI Neo2 in 32 and 64-bit mode.  Moved back to Pi2/3 in 32-bit for
stock default for this config set so that a user can just "go" with
a very obtainable board type.  NanoPI's stuff is nice, but it's
slightly less obtainable right now- while I have it...I've got to
mostly think about prospective end-users since even though it's my
sandbox, it IS intended for other people to grab up and just use.