Trimmed out crap.
Some of this was experiments that just didn't work.

Some of it was things that worked before I moved to tip (Since there's
this extremely disturbing tendency to forgive people checking in
**BROKEN** crap into master because of the way Poky/OpenEmbedded
does their Config Management...)  So, to have an even chance
at something working RIGHT in this for NanoPI Neo2's, etc.
I am clearing the stuff I know I don't need (AND is **BROKEN**)
out of the layer sets for a build.