Pulling meta-kodi out for now...
Found out why I was having a problem with everything.  meta-kodi was
explicitly putting in glamor into xserver-xorg when meta-tegra was
legititmately stripping it out for BSP reasons.  One should not have
a feature forced on like that when it's for something like basically
an application.  Will review what might need to be done to be able
to ultimately build the layer's recipes.  Might need to be peeling
the metadata out, might need to fork the metadata layer.  As it
stands, it has to come out to work correctly for X11 on my current
stuff.